+++ packages/core/app/components/sections/Navbar/Navbar.tsx
@@ -1,5 +1,5 @@
 import Section from '../../../../src/components/sections/Section'
-import { getOverridableSection } from '../../../../src/sdk/overrides/getOverriddenSection'
+import { getOverridableSection } from '../../../sdk/overrides/getOverriddenSection'
 
 import styles from './section.module.scss'
