+++ packages/core/app/components/sections/RegionBar/OverriddenDefaultRegionBar.ts
@@ -1,5 +1,5 @@
 import { override } from 'src/customizations/src/components/overrides/RegionBar'
-import { getOverriddenSection } from 'src/sdk/overrides/getOverriddenSection'
+import { getOverriddenSection } from 'app/sdk/overrides/getOverriddenSection'
 import type { SectionOverrideDefinitionV1 } from 'src/typings/overridesDefinition'
 import RegionBar from '.'
