+++ packages/core/app/fs-next-update/page.tsx
@@ -6,7 +6,7 @@ import { getDynamicContent } from 'src/utils/dynamicContent'
 import storeConfig from '../../faststore.config'
 import { ComponentType } from 'react'
 import { OverriddenDefaultHero as Hero } from 'app/components/sections/Hero/OverriddenDefaultHero'
-import PageProvider from 'src/sdk/overrides/PageProvider'
+import PageProvider from 'app/sdk/overrides/PageProvider'
 import RenderSections from 'app/components/cms/RenderSections'
 
 /* A list of components that can be used in the CMS. */
