+++ packages/core/app/sdk/overrides/overrides.ts
@@ -3,9 +3,9 @@ import type {
   ComponentOverrideDefinition,
   OverriddenComponents,
   SectionOverrideDefinitionV1,
-} from '../../typings/overridesDefinition'
+} from 'src/typings/overridesDefinition'
 
-import type { SectionsOverrides } from '../../typings/overrides'
+import type { SectionsOverrides } from 'src/typings/overrides'
 
 /* TODO: Fix typescript errors. It is necessary to further investigate the typing 
 issues of this file. Error lines have been marked with the @ts-ignore comment*/
