+++ packages/core/src/components/search/Filter/Filter.tsx
@@ -4,7 +4,7 @@ import { Suspense } from 'react'
 
 import { gql } from '@generated'
 import { ProductGalleryProps } from 'src/components/ui/ProductGallery/ProductGallery'
-import { useOverrideComponents } from 'src/sdk/overrides/OverrideContext'
+import { useOverrideComponents } from 'app/sdk/overrides/OverrideContext'
 import { useFilter } from 'src/sdk/search/useFilter'
 
 interface Props {
