+++ packages/core/src/components/sections/BannerText/BannerText.tsx
@@ -5,12 +5,12 @@ import {
   BannerTextContentProps as UIBannerTextContentProps,
 } from '@faststore/ui'
 
-import { useOverrideComponents } from '../../../sdk/overrides/OverrideContext'
+import { useOverrideComponents } from '../../../../app/sdk/overrides/OverrideContext'
 import Section from '../Section'
 
 import styles from './section.module.scss'
 import { BannerTextDefaultComponents } from './DefaultComponents'
-import { getOverridableSection } from '../../../sdk/overrides/getOverriddenSection'
+import { getOverridableSection } from '../../../../app/sdk/overrides/getOverriddenSection'
 
 export interface BannerTextProps {
   title: UIBannerTextContentProps['title']
