+++ packages/core/src/components/sections/CrossSellingShelf/OverriddenDefaultCrossSellingShelf.ts
@@ -1,5 +1,5 @@
 import { override } from 'src/customizations/src/components/overrides/CrossSellingShelf'
-import { getOverriddenSection } from 'src/sdk/overrides/getOverriddenSection'
+import { getOverriddenSection } from 'app/sdk/overrides/getOverriddenSection'
 import type { SectionOverrideDefinitionV1 } from 'src/typings/overridesDefinition'
 import CrossSellingShelf from '.'
