+++ packages/core/src/components/sections/Newsletter/OverriddenDefaultNewsletter.ts
@@ -1,5 +1,5 @@
 import { override } from 'src/customizations/src/components/overrides/Newsletter'
-import { getOverriddenSection } from 'src/sdk/overrides/getOverriddenSection'
+import { getOverriddenSection } from 'app/sdk/overrides/getOverriddenSection'
 import type { SectionOverrideDefinitionV1 } from 'src/typings/overridesDefinition'
 import Newsletter from './Newsletter'
