+++ packages/core/src/components/ui/ProductShelf/ProductShelf.tsx
@@ -2,7 +2,7 @@ import { useEffect, useId, useRef } from 'react'
 
 import { useViewItemListEvent } from 'app/sdk/analytics/hooks/useViewItemListEvent'
 import ProductShelfSkeleton from 'src/components/skeletons/ProductShelfSkeleton'
-import { useOverrideComponents } from 'src/sdk/overrides/OverrideContext'
+import { useOverrideComponents } from 'app/sdk/overrides/OverrideContext'
 import { useProductsQuery } from 'src/sdk/product/useProductsQuery'
 import { textToKebabCase } from 'src/utils/utilities'
